formatting and typecasting for Task3 fixed

+++ src/com/tomas/tasks/primitives/Task3.java
@@ -1,11 +1,11 @@
 package com.tomas.tasks.primitives;
 
 public class Task3 {
-    public static void main(String[] args) {
+  public static void main(String[] args) {
     char t = 't';
-    int ascii = t;
+    int ascii = (int) t;
 
-        System.out.println("Character is: " + t);
-        System.out.println("ASCII number is: " + ascii);
-    }
+    System.out.println("Character is: " + t);
+    System.out.println("ASCII number is: " + ascii);
+  }
 }
